Add parser for label pointers
What a mess! At least we can specify parsers as generic functions and so
pass in a reference to the very start of the packet for resolving the
indirect addresses, and given that the `input` reference is a view into
this very same region we can rely on Rust to guarantee that both live
long enough and `start_of_packet` outlives `input`.